Update
A long time overdue, time to commit.
Have added tons of stuff to the project, such as routes for weapons and locations.
These include mods, base-weapons and properties.
The location comes with attributes and notables.
Due to the location-notables I've added enemies to the models.
Serializers for handling json i/o for most of the additions.